Move sharedFunctions to components/
