+++ src/components/sharedFunctions.ts
